Add main features
This patch adds the main features, which is: copy a file from bucket to
another bucket with the required metadata and the given KMS key if it is
a Streamable file.

This patch lacks of unit tests. Besides, some internal methods need
refactor and there is not any documentation in README. All of them
should be handle at #1, #2, #3 and #4 issues.

Signed-off-by: r1cm3d <r1cm3d@protonmail.com>